requirements.txt
Flask==2.3.3
pandas==2.0.3
mysql-connector-python==8.1.0
numpy==1.24.3
Werkzeug==2.3.7
dropbox
